# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion de Gestion Promos\nPOUR UNE FACILITATION DES TESTS, L'AUTHENTIFICATION EST TEMPORAIREMENT DÉSACTIVÉE !",
        "version": "1.0.0",
        "description": "API Gestion-Promos"
    },
    "paths": {
...
        "/api/v1/referentiels": {
            "get": {
                "tags": ["referentiels"],
                "summary": "Lister tous les référentiels actifs",
                "description": "Recupére la liste des référentiels",
                "responses": {
                    "200": {
                        "description": "Liste des référentiels",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": ["referentiels"],
                "summary": "Créer un nouveau référentiel",
                "description": "Endpoint pour créer un nouveau référentiel.",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "codeReferentiel": {
                                        "type": "string",
                                        "description": "Code unique du référentiel"
                                    },
                                    "libelleReferentiel": {
                                        "type": "string",
                                        "description": "Libellé unique du référentiel"
                                    },
                                    "StatutReferentiel": {
                                        "type": "string",
                                        "description": "Statut du référentiel",
                                        "enum": ["Actif", "Inactif", "Archiver"]
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Image du référentiel"
                                    },
                                    "competences": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "nom": {
                                                    "type": "string",
                                                    "description": "Nom de la compétence"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "description": "Type de compétence",
                                                    "enum": [
                                                        "Back-end",
                                                        "Front-end"
                                                    ]
                                                }
                                            },
                                            "required": ["nom", "type"]
                                        }
                                    }
                                },
                                "required": [
                                    "codeReferentiel",
                                    "libelleReferentiel",
                                    "StatutReferentiel"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Référentiel créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "codeReferentiel": {
                                            "type": "string"
                                        },
                                        "libelleReferentiel": {
                                            "type": "string"
                                        },
                                        "StatutReferentiel": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/referentiels/{id}": {
            "patch": {
                "tags": ["referentiels"],
                "summary": "Modifier un référentiel (ajouter/supprimer des compétences ou des modules)",
                "description": "Ce endpoint permet d'ajouter des compétences, d'ajouter des modules à une compétence, ou de supprimer un module d'une compétence existante.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID du référentiel à modifier"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "description": "L'action à effectuer",
                                        "enum": [
                                            "add_competence",
                                            "add_modules",
                                            "remove_module"
                                        ]
                                    },
                                    "competence_nom": {
                                        "type": "string",
                                        "description": "Nom de la compétence (nécessaire pour ajouter ou supprimer un module)",
                                        "nullable": true
                                    },
                                    "competence": {
                                        "type": "array",
                                        "description": "Détails de la compétence à ajouter (nécessaire pour l'action 'add_competence')",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "nom": {
                                                    "type": "string",
                                                    "description": "Nom de la compétence"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "description": "Description de la compétence",
                                                    "nullable": true
                                                },
                                                "duree_acquisition": {
                                                    "type": "integer",
                                                    "description": "Durée d'acquisition de la compétence en heures",
                                                    "nullable": true
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "description": "Type de la compétence",
                                                    "enum": [
                                                        "Back-end",
                                                        "Front-end"
                                                    ]
                                                }
                                            },
                                            "required": ["nom", "type"]
                                        },
                                        "nullable": true
                                    },
                                    "modules": {
                                        "type": "array",
                                        "description": "Liste des modules à ajouter à la compétence (nécessaire pour l'action 'add_modules')",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "nom": {
                                                    "type": "string",
                                                    "description": "Nom du module"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "description": "Description du module",
                                                    "nullable": true
                                                },
                                                "duree_acquisition": {
                                                    "type": "integer",
                                                    "description": "Durée d'acquisition du module en heures",
                                                    "nullable": true
                                                }
                                            },
                                            "required": ["nom"]
                                        },
                                        "nullable": true
                                    },
                                    "module_nom": {
                                        "type": "string",
                                        "description": "Nom du module à supprimer (nécessaire pour l'action 'remove_module')",
                                        "nullable": true
                                    }
                                },
                                "required": ["action"]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Modification du référentiel réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Compétence ajoutée avec succès"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Référentiel non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Referentiel not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Action invalide ou données manquantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid action"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["referentiels"],
                "summary": "Supprimer un référentiel (soft delete)",
                "description": "Ce endpoint permet de supprimer un référentiel par soft delete (la ressource est marquée comme supprimée sans être réellement effacée de la base de données).",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID du référentiel à supprimer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression du référentiel réussie (soft delete)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Referentiel supprimé avec succès"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Référentiel non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Referentiel not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur lors de la suppression",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors de la suppression du référentiel"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/referentiels/archive/referentiel": {
            "get": {
                "tags": ["referentiels"],
                "summary": "Lister les référentiels supprimés (archivés)",
                "description": "Ce endpoint permet de lister tous les référentiels qui ont été supprimés (ceux marqués avec le champ 'deleted_at').",
                "parameters": [
                    {
                        "name": "statut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": ["archived"]
                        },
                        "description": "Optionnel: Filtrer les référentiels par statut, ex: 'archived' pour les archivés"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des référentiels archivés récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "codeReferentiel": {
                                                "type": "string"
                                            },
                                            "libelleReferentiel": {
                                                "type": "string"
                                            },
                                            "StatutReferentiel": {
                                                "type": "string"
                                            },
                                            "deleted_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Aucun référentiel archivé trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Archive Référentiel vide."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}